requirements.txt
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.0
certifi==2026.1.4
charset-normalizer==3.4.4
crispy-bootstrap5==2025.6
distro==1.9.0
Django==5.2.10
django-crispy-forms==2.5
django-nested-admin==4.1.6
django-widget-tweaks==1.5.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.12.0
openai==2.15.0
pillow==12.1.0
pydantic==2.12.5
pydantic_core==2.41.5
pypdf==6.6.0
python-dotenv==1.2.1
python-monkey-business==1.1.0
reportlab==4.4.9
requests==2.32.5
sniffio==1.3.1
sqlparse==0.5.5
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
django-allauth==65.14.0

# Production Dependencies
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0
psycopg2-binary==2.9.9
